Understand thrust major calculation
The thrust major axis was often found to not be orthogonal to the thrust
axis, as it should be by construction.  This was because of numerical
precision.  The set of input vectors constructed orthogonal to the
thrust axis in the 3d case cannot be constructed to be truly orthogonal.
After this operation, there remains a component of these vectors that is
parallel to the thrust axis.  This component grows in the calculation of
thrust.

To temper this parallel component, it is subtracted from each candidate
thrust axis if a flag indicating that it is the thrust major axis being
calculated is true.